MeshTweaker cannot find numpy module For some reason, when Tweaker-3 is run by autoslice.py as a subprocess (the Tweaker CLI tool), the MeshTweaker class throws exceptions about not finding numpy. I wonder if this has something to do with Tweaker being run as a subprocess (and therefore has  __name__ == '__main__' true), which somehow interferes with the venv set up by printer-manager. It's still all very confusing, but reading https://python-notes.curiousefficiency.org/en/latest/python_concepts/import_traps.html gives me a clue that the way the current implementation attempts to do things causes some sort of issue related to imports...
As I recall, I don't have to do anything like this when Tweaker is run separated from autoslicer, the way things currently are in develop. Here, Tweaker is just run as a subprocess by autoslicer, by running it from the directory specified in the autoslicer (or was it printer-manager) config files.

I think the actual solution here is to either fully implement Tweaker as a submodule in the project, or remove it from the project and run it as a separate process (maybe using that Docker that Tweaker has?). Either way, it seems this mixed approach (both a submodule, AND being run as subprocess) is causing all sorts of issues...